Lower the max duration of batch steps
We are running into issues where the load balancer has a max execution limit of 120 seconds and is thus causing the migration to fail on large sites.  This lowers the max step time to 60 seconds, and the time limit for requests should also be 60 seconds, so the max of 120 seconds should never be reached.